+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+*** 0.0.8
+- Don't cache if DEBUG env var is set
+- Add default `layout`
+- Add `layout` option, effectively the inverse of partial rendering
+- Add `sep` option
+
 *** 0.0.7
 - Improve error stack generation